+++ core/pubnub_netcore.c
@@ -24,6 +24,9 @@
 #if PUBNUB_USE_GRANT_TOKEN_API
 #include "core/pbcc_grant_token_api.h"
 #endif
+#if PUBNUB_USE_REVOKE_TOKEN_API
+#include "core/pbcc_revoke_token_api.h"
+#endif
 #include "core/pubnub_proxy_core.h"
 
 #include <string.h>
@@ -315,6 +318,9 @@ static PFpbcc_parse_response_T m_aParseResponse[] = { dont_parse,
 #if PUBNUB_USE_GRANT_TOKEN_API
     , pbcc_parse_grant_token_api_response /* PBTT_GRANT_TOKEN */
 #endif /* PUBNUB_USE_GRANT_TOKEN_API */
+#if PUBNUB_USE_REVOKE_TOKEN_API
+    , pbcc_parse_revoke_token_response /* PBTT_REVOKE_TOKEN */
+#endif /* PUBNUB_USE_REVOKE_TOKEN_API */
 #endif /* PUBNUB_ONLY_PUBSUB_API */
 };